requirements.txt
jinja2
pyyaml
logging